requirements.txt
# Deep Learning
torch==2.3.1
torchvision==0.18.1

# Computer Vision
opencv-python==4.8.1.78
scikit-image==0.23.2
albumentations==1.3.1

# Machine Learning
xgboost==2.0.0
scikit-learn==1.4.2


# Image Processing
Pillow==10.2.0
imageio==2.34.0

# Data Processing
numpy==1.26.4
pandas 

# Visualization
matplotlib==3.8.4
seaborn==0.13.2

# Progress bars
tqdm==4.66.4

# Configuration
pyyaml==6.0.2
